+++ wisp/wisp.py
@@ -31,7 +31,17 @@ import ConfigParser
 
 from .calibration import Calibration, apply_calibration, generate_tables
 from .flagging import auto_flag, manual_flag, preliminary_flagging
-from .plots import visibility_plots, plotcal_plots
+from .calibration_plots import visibility_plots, plotcal_plots
+
+from .imaging import Imaging
+from .mfs_images import (
+    mfs_dirty_cont,
+    mfs_clean_cont,
+    mfs_dirty_spws,
+    mfs_clean_spws,
+)
+from .channel_images import channel_dirty_spws, channel_clean_spws
+from .imaging_plots import contplot, lineplot
 
 import __main__ as casa
 
@@ -168,6 +178,165 @@ def calibrate(
             manual_flag(cal, "science")
         elif answer == "9":
             cal.split_fields()
+        elif answer.lower() == "q" or answer.lower() == "quit":
+            break
+        else:
+            print("Input not recognized.")
+        if auto_ind >= len(auto_items):
+            break
+
+
+def imaging(
+    vis,
+    field,
+    config_file,
+    outdir=".",
+    stokes="I",
+    spws="",
+    uvrange="",
+    uvtaper=False,
+    outertaper="",
+    interactive=False,
+    savemodel=None,
+    parallel=False,
+    auto="",
+):
+    """
+    Generate and clean images
+
+    Inputs:
+      vis :: string
+        The measurement set containing all data for field
+      field :: string
+        The field name to image
+      config_file :: string
+        filename of the configuration file for this project
+      outdir :: string
+        The directory where to save the results. This is useful if
+        you plan to make several sets of images (i.e. self-calibration)
+      stokes :: string
+        The Stokes parameters we're imaging. e.g. 'I' or 'IQUV'
+      spws :: string
+        comma-separated list of spws to clean
+        if empty, clean all spws
+      uvrange :: string
+        Selection on UV-range
+      uvtaper :: boolean
+        if True, apply UV tapering
+      outertaper :: string
+        Tapering FWHM
+      interactive :: boolean
+        if True, interactively clean
+      savemodel :: string
+        if not none, save individual MFS images of each spectral
+        window to the model column of the measurement set for
+        self-calibration. This can only be done with stokes='I'.
+        if savemodel == 'light': save the model after lightniter
+        if savemodel == 'clean': save the model after niter
+      parallel :: boolean
+        if True, run parallel TCLEAN
+        N.B. CASA must be started in MPI mode (via mpicasa)
+      auto :: string
+        if not an empty string, it is a comma separated
+        list of menu items to perform, i.e. auto='0,1,4,5,6'
+
+    Returns: Nothing
+    """
+    #
+    # start logger
+    #
+    logger = logging.getLogger("main")
+    #
+    # Check inputs
+    #
+    if not os.path.isdir(vis):
+        logger.critical("Measurement set not found!")
+        raise ValueError("Measurement set not found!")
+    if not os.path.exists(config_file):
+        logger.critical("Configuration file not found")
+        raise ValueError("Configuration file not found!")
+    if savemodel is not None and stokes != "I":
+        logger.critical("Can only save visibility model with Stokes I")
+        raise ValueError("Can only save visibility model with Stokes I")
+    #
+    # load configuration file
+    #
+    config = ConfigParser.ConfigParser()
+    logger.info("Reading configuration file {0}".format(config_file))
+    config.read(config_file)
+    logger.info("Done.")
+    #
+    # Initialize Imaging object
+    #
+    img = Imaging(
+        vis,
+        field,
+        logger,
+        config,
+        outdir=outdir,
+        uvtaper=uvtaper,
+        outertaper=outertaper,
+        spws=spws,
+        uvrange=uvrange,
+        stokes=stokes,
+        savemodel=savemodel,
+        interactive=interactive,
+        parallel=parallel,
+    )
+    #
+    # Prompt the user with a menu for each option, or auto-do them
+    #
+    auto_items = auto.split(",")
+    auto_ind = 0
+    while True:
+        if not auto:
+            print(
+                "0. Dirty image combined continuum spws "
+                "(MFS; multi-term; multi-scale)"
+            )
+            print(
+                "1. Clean combined continuum spws "
+                "(MFS; multi-term; multi-scale)"
+            )
+            print("2. Dirty image each continuum spw (MFS; multi-scale)")
+            print("3. Clean each continuum spw (MFS; multi-scale)")
+            print("4. Dirty image each continuum spw (channel; multi-scale)")
+            print("5. Clean each continuum spw (channel; multi-scale)")
+            print("6. Dirty image each line spw (MFS; multi-scale)")
+            print("7. Clean each line spw (MFS; multi-scale)")
+            print("8. Dirty image each line spw (channel; multi-scale)")
+            print("9. Clean each line spw (channel; multi-scale)")
+            print("10. Generate continuum diagnostic plots")
+            print("11. Generate spectral line diagnostic plots")
+            print("q [quit]")
+            answer = raw_input("> ")
+        else:
+            answer = auto_items[auto_ind]
+            auto_ind += 1
+        if answer == "0":
+            mfs_dirty_cont(img)
+        elif answer == "1":
+            mfs_clean_cont(img)
+        elif answer == "2":
+            mfs_dirty_spws(img, img.cont_spws)
+        elif answer == "3":
+            mfs_clean_spws(img, img.cont_spws, "cont")
+        elif answer == "4":
+            channel_dirty_spws(img, img.cont_spws, "cont")
+        elif answer == "5":
+            channel_clean_spws(img, img.cont_spws, "cont")
+        elif answer == "6":
+            mfs_dirty_spws(img, img.line_spws)
+        elif answer == "7":
+            mfs_clean_spws(img, img.line_spws, "line")
+        elif answer == "8":
+            channel_dirty_spws(img, img.line_spws, "line")
+        elif answer == "9":
+            channel_clean_spws(img, img.line_spws, "line")
+        elif answer == "10":
+            contplot(img)
+        elif answer == "11":
+            lineplot(img)
         elif answer.lower() == "q" or answer.lower() == "quit":
             break
         else:
